[refactor]: remove Swarm from Kagami; introduce refactored `iroha_swarm`
