+++ tools/swarm/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "iroha_swarm"
+
+edition.workspace = true
+version.workspace = true
+authors.workspace = true
+license.workspace = true
+
+[dependencies]
+iroha_crypto.workspace = true
+iroha_data_model.workspace = true
+iroha_primitives.workspace = true
+iroha_config.workspace = true
+color-eyre.workspace = true
+expect-test.workspace = true
+path-absolutize.workspace = true
+pathdiff.workspace = true
+owo-colors = { workspace = true, features = ["supports-colors"] }
+serde = { workspace = true, features = ["derive"] }
+clap = { workspace = true, features = ["derive"] }
+serde_yaml.workspace = true
+serde_json.workspace = true
+derive_more.workspace = true
+inquire.workspace = true
+
